# Project/1.PondPlanning/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-based Village Pond Planning System",
    "description": "Terrain, catchment and runoff analysis for village pond siting. Analysis routes are asynchronous: they return 202 with a job id to poll.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rainfall/statistics": {
      "get": {
        "tags": [
          "rainfall"
        ],
        "summary": "Get Statistics",
        "description": "FR5: statistics computed from the record \u2014 not a raw dump.\n\nThe design figure is ``dependable_75``, the annual rainfall equalled or\nexceeded in 75 % of years. Sizing to the mean produces a pond that fails in\nroughly half of them.",
        "operationId": "get_statistics_api_v1_rainfall_statistics_get",
        "parameters": [
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "description": "Longitude, EPSG:4326",
              "title": "Lon"
            },
            "description": "Longitude, EPSG:4326"
          },
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "description": "Latitude, EPSG:4326",
              "title": "Lat"
            },
            "description": "Latitude, EPSG:4326"
          },
          {
            "name": "years",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 5,
              "description": "Length of record to analyse",
              "default": 20,
              "title": "Years"
            },
            "description": "Length of record to analyse"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RainfallStatistics"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rainfall/series": {
      "get": {
        "tags": [
          "rainfall"
        ],
        "summary": "Get Series",
        "description": "The daily record.\n\nExposed because SCS-CN must be applied per day and then summed; applied to an\nannual total it overestimates runoff two- to three-fold.",
        "operationId": "get_series_api_v1_rainfall_series_get",
        "parameters": [
          {
            "name": "lon",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 180,
              "minimum": -180,
              "description": "Longitude, EPSG:4326",
              "title": "Lon"
            },
            "description": "Longitude, EPSG:4326"
          },
          {
            "name": "lat",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 90,
              "minimum": -90,
              "description": "Latitude, EPSG:4326",
              "title": "Lat"
            },
            "description": "Latitude, EPSG:4326"
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO date, inclusive",
              "title": "Start"
            },
            "description": "ISO date, inclusive"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO date, inclusive",
              "title": "End"
            },
            "description": "ISO date, inclusive"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RainfallSeries"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "List recommendations",
        "description": "Return saved pond recommendations.",
        "operationId": "list_recommendations_api_v1_recommendations_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RecommendationOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{recommendation_id}": {
      "get": {
        "tags": [
          "recommendations"
        ],
        "summary": "Get Recommendation",
        "description": "Return one recommendation.",
        "operationId": "get_recommendation_api_v1_recommendations__recommendation_id__get",
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recommendation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{recommendation_id}/status": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Change Status",
        "description": "Approve or reject a recommendation.\n\nRole-gated in P6: a viewer receives ``403``. Every transition writes an\n``audit_log`` row, and that table is append-only in the database \u2014 the rules\nare already in migration 0001, so the trail cannot be rewritten later.",
        "operationId": "change_status_api_v1_recommendations__recommendation_id__status_post",
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recommendation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StatusChangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecommendationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recommendations/{recommendation_id}/exports": {
      "post": {
        "tags": [
          "recommendations"
        ],
        "summary": "Create Export",
        "description": "Generate a PDF, GeoJSON or CSV export and return a link to it.",
        "operationId": "create_export_api_v1_recommendations__recommendation_id__exports_post",
        "parameters": [
          {
            "name": "recommendation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recommendation Id"
            }
          },
          {
            "name": "export_format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "pdf",
              "title": "Export Format"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportDescriptor"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Poll job status",
        "description": "Return progress for one job.\n\n``stage`` is returned alongside ``progress`` because a percentage alone is\nuninformative during a 25-second pipeline; \"filling sinks\" tells the user\nsomething is happening and roughly what.",
        "operationId": "get_job_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel a job",
        "description": "Request cancellation. Idempotent: cancelling a finished job is not an error.",
        "operationId": "cancel_job_api_v1_jobs__job_id__delete",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{job_id}/result": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Fetch a finished job's result",
        "description": "Return the payload of a succeeded job.\n\nUntyped on purpose: the shape depends on ``kind``. The concrete payloads are\ndocumented under ``/analysis/results/*``, which is what a client should read.",
        "operationId": "get_job_result_api_v1_jobs__job_id__result_get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Job Result Api V1 Jobs  Job Id  Result Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/analysis/results/catchment/{job_id}": {
      "get": {
        "tags": [
          "analysis"
        ],
        "summary": "Catchment Result",
        "description": "FR4 result payload.",
        "operationId": "catchment_result_api_v1_analysis_results_catchment__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatchmentResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExportDescriptor": {
        "properties": {
          "export_id": {
            "type": "string",
            "format": "uuid",
            "title": "Export Id"
          },
          "recommendation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recommendation Id"
          },
          "format": {
            "type": "string",
            "enum": [
              "pdf",
              "geojson",
              "csv"
            ],
            "title": "Format"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "export_id",
          "recommendation_id",
          "format",
          "url",
          "size_bytes",
          "expires_at"
        ],
        "title": "ExportDescriptor",
        "description": "A generated export artifact.\n\nReturns a URL rather than bytes so that a large PDF or GeoJSON does not\noccupy an application worker while it streams, and so the link can be shared."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "enum": [
              "queued",
              "running",
              "succeeded",
              "failed",
              "cancelled"
            ],
            "title": "Status"
          },
          "progress": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Progress"
          },
          "stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stage",
            "description": "Current pipeline stage, for the UI"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProblemDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "result_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Url"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "kind",
          "status",
          "progress",
          "created_at"
        ],
        "title": "JobStatus",
        "description": "Progress of an asynchronous job."
      },
...
      "MonthlyNormal": {
        "properties": {
          "month": {
            "type": "integer",
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Month"
          },
          "mean_rainfall": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "rainy_days": {
            "$ref": "#/components/schemas/QuantityOut"
          }
        },
        "type": "object",
        "required": [
          "month",
          "mean_rainfall",
          "rainy_days"
        ],
        "title": "MonthlyNormal",
        "description": "Long-term mean for one calendar month."
      },
      "Page_RecommendationOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RecommendationOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[RecommendationOut]"
      },
...
      "RainfallSeries": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "station_or_grid": {
            "type": "string",
            "title": "Station Or Grid"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "start": {
            "type": "string",
            "format": "date",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "format": "date",
            "title": "End"
          },
          "days": {
            "type": "integer",
            "title": "Days"
          },
          "series": {
            "items": {
              "$ref": "#/components/schemas/RainfallDaily"
            },
            "type": "array",
            "title": "Series"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "source",
          "station_or_grid",
          "latitude",
          "longitude",
          "start",
          "end",
          "days",
          "series"
        ],
        "title": "RainfallSeries",
        "description": "The daily record backing every rainfall statistic."
      },
      "RainfallStatistics": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "years_of_record": {
            "type": "integer",
            "title": "Years Of Record"
          },
          "start_year": {
            "type": "integer",
            "title": "Start Year"
          },
          "end_year": {
            "type": "integer",
            "title": "End Year"
          },
          "mean_annual": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "median_annual": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "dependable_75": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "coefficient_of_variation": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "monsoon_share": {
            "$ref": "#/components/schemas/QuantityOut",
            "description": "Fraction of annual rainfall in Jun-Sep"
          },
          "max_daily_recorded": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "rainy_days_mean": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "monthly_normals": {
            "items": {
              "$ref": "#/components/schemas/MonthlyNormal"
            },
            "type": "array",
            "title": "Monthly Normals"
          },
          "data_completeness": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "fallback_used": {
            "type": "string",
            "enum": [
              "none",
              "cache",
              "secondary_provider"
            ],
            "title": "Fallback Used",
            "description": "How the data was obtained. 'cache' means the live API was unreachable.",
            "default": "none"
          },
          "attribution": {
            "type": "string",
            "title": "Attribution"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/ResultWarning"
            },
            "type": "array",
            "title": "Warnings"
          }
        },
        "type": "object",
        "required": [
          "source",
          "years_of_record",
          "start_year",
          "end_year",
          "mean_annual",
          "median_annual",
          "dependable_75",
          "coefficient_of_variation",
          "monsoon_share",
          "max_daily_recorded",
          "rainy_days_mean",
          "monthly_normals",
          "data_completeness",
          "attribution"
        ],
        "title": "RainfallStatistics",
        "description": "FR5: statistics computed from the record, not a raw dump.\n\n``dependable_75`` is the design figure: the annual rainfall equalled or\nexceeded in 75 % of years. Designing to the mean produces a pond that fails\nroughly half the time, which is why practice uses the 75 % dependable value."
      },
...
      "RecommendationOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "village_id": {
            "type": "string",
            "format": "uuid",
            "title": "Village Id"
          },
          "village_name": {
            "type": "string",
            "title": "Village Name"
          },
          "location": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Location",
            "description": "[lon, lat] in EPSG:4326"
          },
          "catchment_area": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "gross_storage": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "depth": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "indicative_cost": {
            "$ref": "#/components/schemas/QuantityOut"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "moderate",
              "high"
            ],
            "title": "Confidence"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "approved",
              "rejected"
            ],
            "title": "Status"
          },
          "created_by": {
            "type": "string",
            "title": "Created By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "village_id",
          "village_name",
          "location",
          "catchment_area",
          "gross_storage",
          "depth",
          "indicative_cost",
          "confidence",
          "status",
          "created_by",
          "created_at",
          "updated_at"
        ],
        "title": "RecommendationOut",
        "description": "A saved pond recommendation, in its current lifecycle state."
      },
      "ResultWarning": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "severity": {
            "type": "string",
            "enum": [
              "info",
              "caution",
              "critical"
            ],
            "title": "Severity",
            "default": "caution"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ResultWarning",
        "description": "A caveat attached to an otherwise successful result.\n\nWarnings are how this system avoids presenting a number more confidently than\nit deserves \u2014 \"the source DEM is 30 m, so relief below ~5 m is interpolated,\nnot measured\" belongs next to the storage estimate, not buried in a report."
      },
...
      "StatusChangeRequest": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "submitted",
              "approved",
              "rejected"
            ],
            "title": "Status"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "status",
          "reason"
        ],
        "title": "StatusChangeRequest",
        "description": "Approve or reject a recommendation.\n\nRole-gated: a viewer receives ``403``. Every transition writes an\n``audit_log`` row, which is append-only in the database \u2014 both are G6 exit\ncriteria, and the reason field is what makes the trail worth reading."
      },
...
    }
  }
}